Update LinkedList.js
There was an issue in the map refreshPoints method in the condition where we check if the length of the linked list is equal to the max length and then remove the head if it is. I had used a <= condition mistakenly which meant that the LinkedList would always have one element. Used === and fixed it.